Move module and tests to appropriate places
