+++ b2sdk/__init__.py
@@ -38,7 +38,6 @@ except ImportError:  # Python 2.6
 
 logging.getLogger(__name__).addHandler(NullHandler())
 
-
 import b2sdk.version
 __version__ = b2sdk.version.VERSION
 assert __version__  # PEP-0396
